初回起動時位置情報が取れないバグを修正
